Hypothetical worlds and config based values
Reworks a great number of things, this is not going to be an adaquate commit message. The big ticket items are improvements to allow hypothetical worlds branched off of the main world (useful for animating changes on top of the main world state as well as for determining likely results of actions) and improvements to loading information from configuration.

+++ src/main/scala/arx/graphics/text/TextLayouter.scala
@@ -132,7 +132,7 @@ class TextLayouter protected[text](val font : TBitmappedFont, val scale : Float)
 					jumpIfNecessary(width)
 					maxSpacingThisLine = height // ensure that we don't collide with this image on the next line
 
-					addRect(Rectf(area.x + x,area.y + y + (lineHeight - height) / 2, width, height))
+					addRect(Rectf(area.x + x,(area.y + y + (lineHeight - height) / 2).toInt, width, height))
 
 					x += width + minSpacing
 					maxX = math.max(maxX,x)
